[Decomp][Spell] Cohesion split: extract SpellCheckTarget handlers to SpellCheckTarget.cpp
