+++ src/game/WorldHandlers/Spell.cpp
@@ -7146,206 +7146,8 @@ void Spell::UpdatePointers()
     m_targets.Update(m_caster);
 }
 
-/**
- * @brief Checks whether a target matches the spell's creature type restrictions.
- *
- * @param target The target being validated.
- * @return True if the target type is allowed; otherwise, false.
- */
-bool Spell::CheckTargetCreatureType(Unit* target) const
-{
-    uint32 spellCreatureTargetMask = m_spellInfo->TargetCreatureType;
 
-    // Curse of Doom : not find another way to fix spell target check :/
-    if (m_spellInfo->IsFitToFamily(SPELLFAMILY_WARLOCK, UI64LIT(0x0000000200000000)))
-    {
-        // not allow cast at player
-        if (target->GetTypeId() == TYPEID_PLAYER)
-        {
-            return false;
-        }
 
-        spellCreatureTargetMask = 0x7FF;
-    }
-
-    // Dismiss Pet, Taming Lesson and Control Robot skipped
-    if (m_spellInfo->Id == 2641 || m_spellInfo->Id == 23356 || m_spellInfo->Id == 30009)
-    {
-        spellCreatureTargetMask =  0;
-    }
-
-    if (spellCreatureTargetMask)
-    {
-        uint32 TargetCreatureType = target->GetCreatureTypeMask();
-
-        return !TargetCreatureType || (spellCreatureTargetMask & TargetCreatureType);
-    }
-    return true;
-}
-
-/**
- * @brief Gets the current spell container slot used by this spell.
- *
- * @return The current spell container type.
- */
-CurrentSpellTypes Spell::GetCurrentContainer()
-{
-    if (IsNextMeleeSwingSpell())
-    {
-        return (CURRENT_MELEE_SPELL);
-    }
-    else if (IsAutoRepeat())
-    {
-        return (CURRENT_AUTOREPEAT_SPELL);
-    }
-    else if (IsChanneledSpell(m_spellInfo))
-    {
-        return (CURRENT_CHANNELED_SPELL);
-    }
-    else
-    {
-        return (CURRENT_GENERIC_SPELL);
-    }
-}
-
-/**
- * @brief Validates whether a candidate target is acceptable for a specific effect.
- *
- * @param target The target being checked.
- * @param eff The effect index being validated.
- * @return True if the target is valid for the effect; otherwise, false.
- */
-bool Spell::CheckTarget(Unit* target, SpellEffectIndex eff)
-{
-    // Check targets for creature type mask and remove not appropriate (skip explicit self target case, maybe need other explicit targets)
-    if (m_spellInfo->EffectImplicitTargetA[eff] != TARGET_SELF)
-    {
-        if (!CheckTargetCreatureType(target))
-        {
-            return false;
-        }
-    }
-
-    // Check targets for not_selectable unit flag and remove
-    // A player can cast spells on his pet (or other controlled unit) though in any state
-    if (target != m_caster && target->GetCharmerOrOwnerGuid() != m_caster->GetObjectGuid())
-    {
-        // any unattackable target skipped
-        if (target->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE))
-        {
-            return false;
-        }
-
-        // unselectable targets skipped in all cases except TARGET_SCRIPT targeting
-        // in case TARGET_SCRIPT target selected by server always and can't be cheated
-        if ((!m_IsTriggeredSpell || target != m_targets.getUnitTarget()) &&
-            target->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE) &&
-            m_spellInfo->EffectImplicitTargetA[eff] != TARGET_SCRIPT &&
-            m_spellInfo->EffectImplicitTargetB[eff] != TARGET_SCRIPT &&
-            m_spellInfo->EffectImplicitTargetA[eff] != TARGET_AREAEFFECT_INSTANT &&
-            m_spellInfo->EffectImplicitTargetB[eff] != TARGET_AREAEFFECT_INSTANT &&
-            m_spellInfo->EffectImplicitTargetA[eff] != TARGET_AREAEFFECT_CUSTOM &&
-            m_spellInfo->EffectImplicitTargetB[eff] != TARGET_AREAEFFECT_CUSTOM &&
-            m_spellInfo->EffectImplicitTargetA[eff] != TARGET_NARROW_FRONTAL_CONE &&
-            m_spellInfo->EffectImplicitTargetB[eff] != TARGET_NARROW_FRONTAL_CONE)
-        {
-            return false;
-        }
-    }
-
-    // Check player targets and remove if in GM mode or GM invisibility (for not self casting case)
-    if (target != m_caster && target->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (((Player*)target)->GetVisibility() == VISIBILITY_OFF)
-        {
-            return false;
-        }
-
-        if (((Player*)target)->isGameMaster() && !IsPositiveSpell(m_spellInfo->Id))
-        {
-            return false;
-        }
-    }
-
-    // Check targets for LOS visibility (except spells without range limitations )
-    if (!DisableMgr::IsDisabledFor(DISABLE_TYPE_SPELL, m_spellInfo->Id, NULL, SPELL_DISABLE_LOS))
-    {
-        switch (m_spellInfo->Effect[eff])
-        {
-            case SPELL_EFFECT_SUMMON_PLAYER:                    // from anywhere
-                break;
-            case SPELL_EFFECT_DUMMY:
-                if (m_spellInfo->Id != 20577)                   // Cannibalize
-                {
-                    break;
-                }
-                // fall through
-            case SPELL_EFFECT_RESURRECT_NEW:
-                // player far away, maybe his corpse near?
-                if (target != m_caster && !target->IsWithinLOSInMap(m_caster))
-                {
-                    if (!m_targets.getCorpseTargetGuid())
-                    {
-                        return false;
-                    }
-
-                    Corpse* corpse = m_caster->GetMap()->GetCorpse(m_targets.getCorpseTargetGuid());
-                    if (!corpse)
-                    {
-                        return false;
-                    }
-
-                    if (target->GetObjectGuid() != corpse->GetOwnerGuid())
-                    {
-                        return false;
-                    }
-
-                    if (!corpse->IsWithinLOSInMap(m_caster))
-                    {
-                        return false;
-                    }
-                }
-
-                // all ok by some way or another, skip normal check
-                break;
-            default:                                            // normal case
-            {
-                // Get GO cast coordinates if original caster -> GO
-                if (target != m_caster)
-                {
-                    if (WorldObject* caster = GetCastingObject())
-                    {
-                        if (!target->IsWithinLOSInMap(caster))
-                        {
-                            return false;
-                        }
-                    }
-                }
-                break;
-            }
-        }
-    }
-
-    if (target->GetTypeId() != TYPEID_PLAYER && m_spellInfo->HasAttribute(SPELL_ATTR_EX3_TARGET_ONLY_PLAYER)
-        && m_spellInfo->EffectImplicitTargetA[eff] != TARGET_SCRIPT && m_spellInfo->EffectImplicitTargetA[eff] != TARGET_SELF)
-    {
-        return false;
-    }
-
-    switch (m_spellInfo->Id)
-    {
-        case 37433:                                         // Spout (The Lurker Below), only players affected if its not in water
-            if (target->GetTypeId() != TYPEID_PLAYER || target->IsInWater())
-            {
-                return false;
-            }
-            break;
-        default:
-            break;
-    }
-
-    return true;
-}
 
 /**
  * @brief Checks whether this spell cast should produce client-visible packets.
